+++ demo-ng/src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { NativeScriptModule } from 'nativescript-angular/nativescript.module';
-import { ExamplesModule } from '~/app/examples/examples.module';
-import { HomeModule } from '~/app/home/home.module';
+import { ExamplesModule } from './examples/examples.module';
+import { HomeModule } from './home/home.module';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';