+++ src/cardview/cardview.android.ts
@@ -4,13 +4,13 @@ import { backgroundInternalProperty } from 'tns-core-modules/ui/styling/style-pr
 import { createRippleDrawable, createStateListAnimator, getAttrColor, isPostLollipop, isPostLollipopMR1 } from 'nativescript-material-core/android/utils';
 import { Color } from 'tns-core-modules/color';
 
-let MDCCardView: typeof android.support.design.card.MaterialCardView;
+let MDCCardView: typeof com.google.android.material.card.MaterialCardView;
 
 const DEFAULT_STROKE_VALUE = -1;
 function initMDCCardView() {
     if (!MDCCardView) {
         // if (android.os.Build.VERSION.SDK_INT >= 23) {
-        MDCCardView = android.support.design.card.MaterialCardView;
+        MDCCardView = com.google.android.material.card.MaterialCardView;
         // } else {
         //     initializePreLollipopCardView()
         //     MDCCardView = PreLollipopCardView as any
@@ -18,7 +18,7 @@ function initMDCCardView() {
     }
 }
 
-// interface PreLollipopCardView extends android.support.design.card.MaterialCardView {
+// interface PreLollipopCardView extends com.google.android.material.card.MaterialCardView {
 //     // tslint:disable-next-line:no-misused-new
 //     new (context): PreLollipopCardView;
 // }
@@ -28,7 +28,7 @@ function initMDCCardView() {
 //     if (PreLollipopCardView) {
 //         return;
 //     }
-//     class PreLollipopCardViewImpl extends android.support.design.card.MaterialCardView {
+//     class PreLollipopCardViewImpl extends com.google.android.material.card.MaterialCardView {
 //         constructor(context) {
 //             super(context);
 //             return global.__native(this);
@@ -219,15 +219,15 @@ function initializeOutlineProvider() {
 }
 
 export class CardView extends CardViewBase {
-    nativeViewProtected: android.support.design.card.MaterialCardView;
+    nativeViewProtected: com.google.android.material.card.MaterialCardView;
 
-    fgDrawable: android.graphics.drawable.GradientDrawable;
-    rippleDrawable: android.graphics.drawable.StateListDrawable | android.graphics.drawable.RippleDrawable;
-    rippleShape: android.graphics.drawable.ShapeDrawable;
+    // fgDrawable: android.graphics.drawable.GradientDrawable;
+    // rippleDrawable: android.graphics.drawable.StateListDrawable | android.graphics.drawable.RippleDrawable;
+    // rippleShape: android.graphics.drawable.ShapeDrawable;
     _strokeWidth = 0;
     _borderRadius = 0;
 
-    get android(): android.support.design.card.MaterialCardView {
+    get android(): com.google.android.material.card.MaterialCardView {
         return this.nativeView;
     }
 
@@ -239,16 +239,16 @@ export class CardView extends CardViewBase {
             createStateListAnimator(this, view);
         }
         view.setClickable(this.isUserInteractionEnabled);
-
+        // view.setUseCompatPadding(true)
         // store the default radius
         this._borderRadius = view.getRadius();
 
         // set the view outline
-        view.setClipToOutline(false);
-        const contentView = view.getChildAt(0) || view;
-        contentView.setClipToOutline(true);
-        contentView.setOutlineProvider(new ViewOutlineProvider(new WeakRef(this)));
-        this.createForegroundDrawable(view, this.style.borderWidth, this.style.borderColor as any);
+        // view.setClipToOutline(false);
+        // const contentView = view.getChildAt(0) || view;
+        // contentView.setClipToOutline(true);
+        // contentView.setOutlineProvider(new ViewOutlineProvider(new WeakRef(this)));
+        // this.createForegroundDrawable(view, this.style.borderWidth, this.style.borderColor as any);
         return view;
     }
 
@@ -256,49 +256,53 @@ export class CardView extends CardViewBase {
         const color = this.style['rippleColor'] ? this.style['rippleColor'] : new Color(getAttrColor(this._context, 'colorControlHighlight'));
         return color.android;
     }
-    createForegroundDrawable(view: android.support.design.card.MaterialCardView, strokeWidth, strokeColor: Color) {
-        this.fgDrawable = new android.graphics.drawable.GradientDrawable();
-        const radius = view.getRadius();
-        this.fgDrawable.setCornerRadius(radius);
+    // createForegroundDrawable(view: com.google.android.material.card.MaterialCardView, strokeWidth, strokeColor: Color) {
+    //     this.fgDrawable = new android.graphics.drawable.GradientDrawable();
+    //     const radius = view.getRadius();
+    //     this.fgDrawable.setCornerRadius(radius);
 
-        // In order to set a stroke, a size and color both need to be set. We default to a zero-width
-        // width size, but won't set a default color. This prevents drawing a stroke that blends in with
-        // the card but that could affect card spacing.
-        if (strokeColor && strokeColor.android !== DEFAULT_STROKE_VALUE) {
-            this.fgDrawable.setStroke(strokeWidth, strokeColor.android);
-        }
+    //     // In order to set a stroke, a size and color both need to be set. We default to a zero-width
+    //     // width size, but won't set a default color. This prevents drawing a stroke that blends in with
+    //     // the card but that could affect card spacing.
+    //     if (strokeColor && strokeColor.android !== DEFAULT_STROKE_VALUE) {
+    //         this.fgDrawable.setStroke(strokeWidth, strokeColor.android);
+    //     }
 
-        this.rippleDrawable = createRippleDrawable(view, this.getCardRippleColor(), radius);
-        this.rippleShape = (this.rippleDrawable as any).rippleShape;
+    //     this.rippleDrawable = createRippleDrawable(view, this.getCardRippleColor(), radius);
+    //     this.rippleShape = (this.rippleDrawable as any).rippleShape;
 
-        const result = Array.create(android.graphics.drawable.Drawable, 2);
-        result[0] = this.rippleDrawable;
-        result[1] = this.fgDrawable;
-        view.setForeground(new android.graphics.drawable.LayerDrawable(result));
-    }
+    //     const result = Array.create(android.graphics.drawable.Drawable, 2);
+    //     result[0] = this.rippleDrawable;
+    //     result[1] = this.fgDrawable;
+    //     view.setForeground(new android.graphics.drawable.LayerDrawable(result));
+    // }
 
-    updateBorderRadius(radius) {
-        this._borderRadius = radius;
-        this.fgDrawable.setCornerRadius(this._borderRadius); // for the foreground drawable
-        const radii = Array.create('float', 8);
-        java.util.Arrays.fill(radii, this._borderRadius);
-        this.rippleShape.setShape(new android.graphics.drawable.shapes.RoundRectShape(radii, null, null)); // for the ripple drawable
-    }
+    // updateBorderRadius(radius) {
+    //     this._borderRadius = radius;
+    //     this.fgDrawable.setCornerRadius(this._borderRadius); // for the foreground drawable
+    //     const radii = Array.create('float', 8);
+    //     java.util.Arrays.fill(radii, this._borderRadius);
+    //     this.rippleShape.setShape(new android.graphics.drawable.shapes.RoundRectShape(radii, null, null)); // for the ripple drawable
+    // }
     [backgroundInternalProperty.setNative](value: any) {
         if (this.nativeViewProtected) {
             if (value instanceof android.graphics.drawable.Drawable) {
                 this.nativeViewProtected.setBackgroundDrawable(value);
             } else {
                 this._strokeWidth = value.borderTopWidth;
+                this.nativeViewProtected.setStrokeWidth(this._strokeWidth);
                 if (value.color) {
                     this.nativeViewProtected.setCardBackgroundColor(value.color.android);
                 }
                 if (value.borderTopColor && value.borderTopColor.android !== DEFAULT_STROKE_VALUE) {
-                    this.fgDrawable.setStroke(this._strokeWidth, value.borderTopColor.android);
+                    this.nativeViewProtected.setStrokeColor(value.borderTopColor.android);
+                    // this.fgDrawable.setStroke(this._strokeWidth, value.borderTopColor.android);
                 }
 
                 if (this._borderRadius !== value.borderTopLeftRadius) {
-                    this.updateBorderRadius(value.borderTopLeftRadius);
+                    this._borderRadius = value.borderTopLeftRadius;
+                    this.nativeViewProtected.setRadius(this._borderRadius);
+                    // this.updateBorderRadius(value.borderTopLeftRadius);
                 }
             }
         }
@@ -308,7 +312,7 @@ export class CardView extends CardViewBase {
         if (!this.nativeViewProtected) {
             return;
         }
-        android.support.v4.view.ViewCompat.setElevation(this.nativeViewProtected, value);
+        androidx.core.view.ViewCompat.setElevation(this.nativeViewProtected, value);
         if (isPostLollipop()) {
             createStateListAnimator(this, this.nativeViewProtected);
         }
@@ -323,10 +327,11 @@ export class CardView extends CardViewBase {
     }
     [rippleColorProperty.setNative](color: Color) {
         const rippleColor = color ? color.android : -1;
-        if (isPostLollipopMR1()) {
-            (this.rippleDrawable as android.graphics.drawable.RippleDrawable).setColor(android.content.res.ColorStateList.valueOf(rippleColor));
-        } else {
-            this.rippleShape.getPaint().setColor(rippleColor);
-        }
+        // if (isPostLollipopMR1()) {
+        //     (this.rippleDrawable as android.graphics.drawable.RippleDrawable).setColor(android.content.res.ColorStateList.valueOf(rippleColor));
+        // } else {
+        //     this.rippleShape.getPaint().setColor(rippleColor);
+        // }
+        (this.nativeViewProtected as any).setRippleColor(android.content.res.ColorStateList.valueOf(rippleColor));
     }
 }