+++ src/snackbar/snackbar-common.ts
@@ -4,17 +4,17 @@ export interface SnackBarOptions {
     /**
      * The action button text of the snackbar.
      */
-    actionText: string;
+    actionText?: string;
 
     /**
      * The text of the snackbar.
      */
-    snackText: string;
+    message: string;
 
     /**
      * Delay to hide the snackbar.
      */
-    hideDelay: number;
+    hideDelay?: number;
 
     /**
      * Action Text Color of the snackbar.
@@ -57,4 +57,44 @@ export enum DismissReasons {
     MANUAL = 'Manual',
     CONSECUTIVE = 'Consecutive',
     UNKNOWN = 'Unknown'
+}
+
+export abstract class SnackBarBase {
+    _options: SnackBarOptions;
+    constructor(options?: SnackBarOptions) {
+        this._options = options;
+    }
+
+    public simple(message: string, textColor?: string, backgroundColor?: string, maxLines?: number, isRTL?: boolean, view?: View): Promise<any> {
+        return this.showSnack({
+            message,
+            textColor,
+            backgroundColor,
+            maxLines,
+            isRTL,
+            view
+        });
+    }
+
+    public action(options: SnackBarOptions) {
+        return this.showSnack(options);
+    }
+
+    public showSnack(options: SnackBarOptions) {
+        // if (!options) {
+        //     options = this._options;
+        // }
+        this._options = options;
+        // return new Promise((resolve, reject) => {
+        // try {
+        // this.initSnack(options, resolve);
+        return this.show();
+        // } catch (ex) {
+        // reject(ex);
+        // }
+        // });
+    }
+    abstract show();
+    abstract dismiss();
+    abstract initSnack(options: SnackBarOptions, resolve?: Function);
 }