+++ src/textfield/textfield.android.ts
@@ -1,12 +1,12 @@
 import { TextFieldBase } from './textfield.common';
-import { hintProperty, placeholderColorProperty } from 'tns-core-modules/ui/editable-text-base/editable-text-base';
+import { backgroundInternalProperty, borderBottomLeftRadiusProperty, hintProperty, placeholderColorProperty } from 'tns-core-modules/ui/editable-text-base/editable-text-base';
 
-import * as application from 'application';
-import { errorColorProperty, errorProperty, floatingProperty, helperProperty, highlightColorProperty, maxLengthProperty } from './textfield_cssproperties';
+import { errorColorProperty, errorProperty, floatingColorProperty, floatingProperty, helperProperty, maxLengthProperty, strokeColorProperty } from './textfield_cssproperties';
 import { Color } from 'tns-core-modules/color';
 import { ad } from 'tns-core-modules/utils/utils';
-import { getFocusedColorStateList, handleClearFocus } from 'nativescript-material-core/android/utils';
+import { getFocusedColorStateList, getLayout, handleClearFocus, stateSets } from 'nativescript-material-core/android/utils';
 import { Background } from 'tns-core-modules/ui/styling/background';
+import { layout } from 'tns-core-modules/utils/utils';
 
 declare module 'tns-core-modules/ui/text-field/text-field' {
     interface TextField {
@@ -14,12 +14,7 @@ declare module 'tns-core-modules/ui/text-field/text-field' {
     }
 }
 
-function getLayout(id: string) {
-    const context: android.content.Context = application.android.context;
-    return context.getResources().getIdentifier(id, 'layout', context.getPackageName());
-}
-
-interface TextInputEditText extends android.support.design.widget.TextInputEditText {
+interface TextInputEditText extends com.google.android.material.textfield.TextInputEditText {
     // tslint:disable-next-line:no-misused-new
     new (context): TextInputEditText;
     owner: WeakRef<TextField>;
@@ -30,8 +25,8 @@ export function initTextInputEditText() {
     if (TextInputEditText) {
         return;
     }
-    @JavaProxy('org.nativescript.widgets.TextInputEditText')
-    class TextInputEditTextImpl extends android.support.design.widget.TextInputEditText {
+    @JavaProxy('org.nativescript.material.TextInputEditText')
+    class TextInputEditTextImpl extends com.google.android.material.textfield.TextInputEditText {
         public owner: WeakRef<TextField>;
         constructor(context: android.content.Context) {
             super(context);
@@ -54,16 +49,26 @@ export function initTextInputEditText() {
     TextInputEditText = TextInputEditTextImpl as any;
 }
 
-export class TextField extends TextFieldBase {
-    editText: android.support.design.widget.TextInputEditText;
-    layoutView: android.support.design.widget.TextInputLayout;
+export function getDefaultHintTextColorStateList(pressedColor: number, color = 1627389952) {
+    const states = Array.create('[I', 2);
+    states[0] = stateSets.FOCUSED_STATE_SET;
+    states[1] = Array.create('int', 0);
+    const colors = Array.create('int', 2);
+    colors[0] = pressedColor;
+    colors[1] = color;
+    return new android.content.res.ColorStateList(states, colors);
+}
 
-    // nativeViewProtected: android.support.design.widget.TextInputLayout;
+export class TextField extends TextFieldBase {
+    editText: com.google.android.material.textfield.TextInputEditText;
+    layoutView: com.google.android.material.textfield.TextInputLayout;
+
+    // nativeViewProtected: com.google.android.material.textfield.TextInputLayout;
     constructor() {
         super();
     }
     get nativeTextViewProtected() {
-        return this.editText as android.support.design.widget.TextInputEditText;
+        return this.editText as com.google.android.material.textfield.TextInputEditText;
     }
 
     drawingBackground = false;
@@ -76,34 +81,55 @@ export class TextField extends TextFieldBase {
 
     public createNativeView() {
         let layoutIdName = 'material_text_field';
-        if (this.style['variant'] === 'filled') {
+        if (this.variant === 'filled') {
             layoutIdName = 'material_text_field_filled';
-        } else if (this.style['variant'] === 'outline') {
+        } else if (this.variant === 'outline') {
             layoutIdName = 'material_text_field_outline';
         }
         const layoutId = getLayout(layoutIdName);
 
-        let layout: android.support.design.widget.TextInputLayout;
+        let layoutView: com.google.android.material.textfield.TextInputLayout;
         let editText: TextInputEditText;
 
         initTextInputEditText();
         if (layoutId !== 0) {
-            layout = this.layoutView = android.view.LayoutInflater.from(this._context).inflate(layoutId, null, false) as android.support.design.widget.TextInputLayout;
-            editText = this.editText = (layout.getChildAt(0) as android.widget.FrameLayout).getChildAt(0) as TextInputEditText;
+            layoutView = this.layoutView = android.view.LayoutInflater.from(this._context).inflate(layoutId, null, false) as com.google.android.material.textfield.TextInputLayout;
+            editText = this.editText = (layoutView.getChildAt(0) as android.widget.FrameLayout).getChildAt(0) as TextInputEditText;
         } else {
-            layout = this.layoutView = new android.support.design.widget.TextInputLayout(this._context);
-            editText = this.editText = new TextInputEditText(layout.getContext());
+            layoutView = this.layoutView = new com.google.android.material.textfield.TextInputLayout(this._context);
+            editText = this.editText = new TextInputEditText(layoutView.getContext());
             editText.setLayoutParams(new android.widget.LinearLayout.LayoutParams(android.widget.FrameLayout.LayoutParams.MATCH_PARENT, android.widget.FrameLayout.LayoutParams.WRAP_CONTENT));
-            layout.addView(editText);
+            layoutView.addView(editText);
         }
+        // in com.google.material the default style is boxed!
+        if (layoutIdName === 'material_text_field') {
+            layoutView.setBoxBackgroundColor(android.graphics.Color.TRANSPARENT);
+            editText.setBackgroundColor(android.graphics.Color.TRANSPARENT);
+            // editText.setPadding(0, layout.toDevicePixels(20), 0, layout.toDevicePixels(6));
+        }
+        console.log(
+            this.layoutView.getBoxCornerRadiusTopStart(),
+            this.layoutView.getBoxCornerRadiusTopEnd(),
+            this.layoutView.getBoxCornerRadiusBottomStart(),
+            this.layoutView.getBoxCornerRadiusBottomEnd()
+        );
+        // this.style.borderTopLeftRadius = { unit: 'px', value: this.layoutView.getBoxCornerRadiusTopStart() };
+        // this.style.borderTopRightRadius = { unit: 'px', value: this.layoutView.getBoxCornerRadiusTopEnd() };
+        // this.style.borderBottomLeftRadius = { unit: 'px', value: this.layoutView.getBoxCornerRadiusBottomStart() };
+        // this.style.borderBottomRightRadius = { unit: 'px', value: this.layoutView.getBoxCornerRadiusBottomEnd() };
         editText.owner = new WeakRef(this);
         // this.editText.setSupportBackgroundTintList(android.content.res.ColorStateList.valueOf(new Color('orange').android));
-        layout.setFocusableInTouchMode(true); // to prevent focus on view creation
+        layoutView.setFocusableInTouchMode(true); // to prevent focus on view creation
         // layout.setFocusable(true);
         // layout.setAddStatesFromChildren(true);
         // layout.setDescendantFocusability(android.view.ViewGroup.FOCUS_AFTER_DESCENDANTS);
         // layout.setDescendantFocusability(android.view.ViewGroup.FOCUS_AFTER_DESCENDANTS);
-        return layout;
+        return layoutView;
+    }
+
+    [borderBottomLeftRadiusProperty.getDefault]() {
+        console.log('borderBottomLeftRadiusProperty.getDefault', this.layoutView.getBoxCornerRadiusTopStart());
+        return this.layoutView.getBoxCornerRadiusTopStart();
     }
 
     _redrawNativeBackground(value: android.graphics.drawable.Drawable | Background): void {
@@ -122,9 +148,17 @@ export class TextField extends TextFieldBase {
     }
 
     [placeholderColorProperty.setNative](value: Color) {
-        const color = value instanceof Color ? value.android : value;
-        (this.layoutView as any).setDefaultHintTextColor(android.content.res.ColorStateList.valueOf(color));
+        const placeholderColor = value instanceof Color ? value.android : value;
+        const floatingColor = this.floatingColor instanceof Color ? this.floatingColor.android : placeholderColor;
+
+        this.layoutView.setDefaultHintTextColor(getDefaultHintTextColorStateList(floatingColor, placeholderColor));
     }
+    [floatingColorProperty.setNative](value: Color) {
+        const floatingColor = value instanceof Color ? value.android : value;
+        const placeholderColor = this.placeholderColor instanceof Color ? this.placeholderColor.android : undefined;
+        this.layoutView.setDefaultHintTextColor(getDefaultHintTextColorStateList(floatingColor, placeholderColor));
+    }
+
     public requestFocus() {
         if (this.layoutView) {
             // because of setFocusableInTouchMode fix we need this for focus to work
@@ -168,9 +202,32 @@ export class TextField extends TextFieldBase {
         this.layoutView.setHintEnabled(!!value);
     }
 
-    [highlightColorProperty.setNative](value: Color) {
+    [strokeColorProperty.setNative](value: Color) {
         const color = value instanceof Color ? value.android : value;
-        this.editText.setSupportBackgroundTintList(getFocusedColorStateList(color, this.style['variant']));
+        this.layoutView.setBoxStrokeColor(color);
         // this.editText.setBackgroundTintList(android.content.res.ColorStateList.valueOf(color));
     }
+    [backgroundInternalProperty.setNative](value: Background) {
+        if (this.nativeViewProtected) {
+            if (value instanceof android.graphics.drawable.Drawable) {
+                this.nativeViewProtected.setBackgroundDrawable(value);
+            } else {
+                if (value.color) {
+                    this.layoutView.setBoxBackgroundColor(value.color.android);
+                }
+                if (value.borderTopColor) {
+                    // TODO: for now no control over border color. it is an attr
+                    // this.nativeViewProtected.setStrokeColor(value.borderTopColor.android);
+                }
+
+                // TODO: for now no control over borderRadius because we can't define the default value
+                // this.borderTopLeftRadius = value.borderTopLeftRadius;
+                // this.borderTopRightRadius = value.borderTopRightRadius;
+                // this.borderBottomLeftRadius = value.borderBottomLeftRadius;
+                // this.borderBottomRightRadius = value.borderBottomRightRadius;
+                // this.layoutView.setBoxCornerRadii(this.borderTopLeftRadius, this.borderTopRightRadius, this.borderBottomLeftRadius, this.borderBottomRightRadius);
+            }
+        }
+    }
 }
+//