+++ src/textfield/textfield.d.ts
@@ -19,7 +19,7 @@ export type TextFieldProperties = Partial<Pick<TextField, keyof TextField>>;
 export class TextField extends NTextField {
     /*
      * nativeView
-     * @Android : android.support.design.widget.TextInputLayout
+    * @Android : com.google.android.material.textfield.TextInputLayout
      * @iOS : MDCTextField
      */
     nativeViewProtected: any;