+++ packages/nativescript-material-bottomsheet/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.1](https://github.com/Akylas/nativescript-material-components/compare/v2.3.0...v2.3.1) (2019-07-12)
+
+**Note:** Version bump only for package nativescript-material-bottomsheet
+
+
+
+
+
+# [2.3.0](https://github.com/Akylas/nativescript-material-components/compare/v2.2.17...v2.3.0) (2019-07-12)
+
+
+### Features
+
+* migrate to androidx ([1dc9322](https://github.com/Akylas/nativescript-material-components/commit/1dc9322))
+* **angular-support:** add angular support for bottom sheet ([2b25457](https://github.com/Akylas/nativescript-material-components/commit/2b25457))
+* **angular-support:** complete support for bottom sheet ([1554311](https://github.com/Akylas/nativescript-material-components/commit/1554311))
+
+
+
+
+
 ## [2.2.17](https://github.com/Akylas/nativescript-material-components/compare/v2.2.16...v2.2.17) (2019-06-11)
 
 **Note:** Version bump only for package nativescript-material-bottomsheet